Fix examples; minor renaming
