+++ examples/avro_specific_consumer_example/avro_specific_consumer_example.go
@@ -24,7 +24,8 @@ import (
 	"fmt"
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/confluentinc/confluent-kafka-go/schemaregistry"
-	"github.com/confluentinc/confluent-kafka-go/serde"
+	"github.com/confluentinc/confluent-kafka-go/schemaregistry/serde"
+	"github.com/confluentinc/confluent-kafka-go/schemaregistry/serde/avro"
 	"os"
 	"os/signal"
 	"syscall"
@@ -64,7 +65,7 @@ func main() {
 
 	fmt.Printf("Created Consumer %v\n", c)
 
-	deser, err := serde.NewSpecificAvroDeserializer(&schemaregistry.ConfigMap{
+	deser, err := avro.NewSpecificAvroDeserializer(&schemaregistry.ConfigMap{
 		"schema.registry.url": url,
 	}, serde.ValueSerde)
